requirements.txt
django
pyvmomi
easysnmp
django-apscheduler
daphne